Order-related classes: Add editing of order status
For the delivery system to work, one important feature is to be able
to update the delivery status.

Let's:
* Add appropriate functionalities to edit order status

Note:
* This is merely to edit the order status, not any other info about
the order. The two features are deliberately separated as they have
different purposes and different usage frequencies, for instance
editing the status will be used a lot more frequently and the other
will only be used in case of errors.